+++ pg_m2tn/models/__init__.py
@@ -1,4 +1,0 @@
-from pg_m2tn.models.loss import FixedWeightedLoss
-from pg_m2tn.models.pg_m2tn import PGM2TN, count_parameters
-
-__all__ = ["PGM2TN", "FixedWeightedLoss", "count_parameters"]