+++ pg_m2tn/utils/__init__.py
@@ -1,3 +1,0 @@
-from pg_m2tn.utils.metrics import compute_metrics, per_dataset_metrics
-
-__all__ = ["compute_metrics", "per_dataset_metrics"]